+++ app/src/main/java/projects/medicationtracker/MedicationNotes.java
@@ -28,7 +28,7 @@ import projects.medicationtracker.Dialogs.AddNoteDialog;
 import projects.medicationtracker.Helpers.DBHelper;
 import projects.medicationtracker.Helpers.TextViewUtils;
 import projects.medicationtracker.Interfaces.IDialogCloseListener;
-import projects.medicationtracker.SimpleClasses.Note;
+import projects.medicationtracker.Models.Note;
 import projects.medicationtracker.Views.StandardCardView;
 
 public class MedicationNotes extends AppCompatActivity implements IDialogCloseListener {
