+++ app/src/main/java/projects/medicationtracker/MyMedications.java
@@ -26,7 +26,7 @@ import java.util.stream.Collectors;
 import kotlin.Pair;
 import projects.medicationtracker.Fragments.MyMedicationsFragment;
 import projects.medicationtracker.Helpers.DBHelper;
-import projects.medicationtracker.SimpleClasses.Medication;
+import projects.medicationtracker.Models.Medication;
 import projects.medicationtracker.Views.StandardCardView;
 
 public class MyMedications extends AppCompatActivity {
